refactor: 拆分mclogs相关功能
